아이템 모드 실행 시 게임프레임에 아이템 표시

+++ src/playscreen/panels/ScorePanel.java
@@ -7,21 +7,21 @@ import java.awt.Font;
 import java.awt.BorderLayout;
 
 public class ScorePanel extends JPanel {
-    private int screenSize;
+    private double screenSize;
     private int score;
     private JLabel scoreLabel;
 
     // Constructor
-    public ScorePanel(int screenSize, int score) {
+    public ScorePanel(double screenSize, int score) {
         this.screenSize = screenSize;
         this.score = score;
 
         this.setLayout(new BorderLayout());
         scoreLabel = new JLabel(String.valueOf(score), SwingConstants.CENTER);
-        scoreLabel.setFont(new Font("Serif", Font.BOLD, 20 * screenSize));
+        scoreLabel.setFont(new Font("Serif", Font.BOLD, (int)(20 * screenSize)));
         this.add(scoreLabel, BorderLayout.CENTER);
 
-        this.setPreferredSize(new java.awt.Dimension(10 * 20 * screenSize, 5 * 20 * screenSize));
+        this.setPreferredSize(new java.awt.Dimension((int)(10 * 20 * screenSize), (int)(5 * 20 * screenSize)));
     }
 
     // 점수 업데이트 메서드
